feat: feitos os services para calculo da SAC e PRICE. Atualizado o Service e finalizada a simulação conforme pedido

+++ src/main/java/com/hackaton/simulacao/api/dto/RequestDTO.java
@@ -16,5 +16,5 @@ import java.math.BigDecimal;
 public class RequestDTO {
 
     private BigDecimal valorDesejado;
-    private Long prazo;
+    private int prazo;
 }
